feat: moved log transport to main package
