+++ electron/main/logger/__mocks__/formatter.mock.ts
