+++ electron/main/logger/__mocks__/transporter.mock.ts
