+++ electron/main/logger/transport/__tests__/file-transporter.test.ts
@@ -1,7 +1,7 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { LogLevel, type LogMessage } from '../../../../common/logger/types.js';
-import type { LogFormatterMock } from '../../__mocks__/log-formatter.mock.js';
-import { LogFormatterMockImpl } from '../../__mocks__/log-formatter.mock.js';
+import type { LogFormatterMock } from '../../__mocks__/formatter.mock.js';
+import { LogFormatterMockImpl } from '../../__mocks__/formatter.mock.js';
 import { FileLogTransporterImpl } from '../file.transporter.js';
 
 const { mockWritable, mockCreateWriteStream } = await vi.hoisted(async () => {
